// types/MapaDeCarregamento.ts
export interface IMapaCarregamento {
    Code: number;
    U_NPedido: number;
    U_Data: string;

    U_CodItem: string;
    U_Lote: string;
    U_Invoice: string;
    U_CodClient: string;
    U_CardName: string;
    U_DescItem: string;
    U_Unidade: string;

    U_Quant: string;
    U_Preco: string;
    U_Total: string;

    U_Fabric: string;
    U_Validade: string;

    U_Cidade: string;
    U_Pais: string;
    U_Bairro: string;
    U_Cep: string;
    U_Rua: string;

    U_Deposito: string;

    U_PPalet: string;
    U_PRef: string;
    U_PLiq: string;
    U_PBru: string;
    U_NPalet: string;

    //LineNum: number;
}